requirements-dev.txt
pytest==5.3.2
pytest-cov==2.8.1
codecov==2.1.13
requests-flask-adapter==0.0.5
requests-http-signature==0.1.0
requests==2.22.0
black==22.12.0
isort==5.11.4
flake8==6.0.0
hypothesis==6.70.1
